% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recursive_utils.R
\name{extract_recursive_priors}
\alias{extract_recursive_priors}
\title{Extract Last State as Priors}
\usage{
extract_recursive_priors(fit, teams)
}
\arguments{
\item{fit}{A CmdStanFit object (either full history or recursive).}

\item{teams}{Character vector of team abbreviations (to ensure alignment).}
}
\value{
A list of vectors (mean/sd) for recursive priors.
}
\description{
Extracts the final week's team strengths and the season's HFA/Alpha
to serve as priors for the next recursive step.
}
