% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{next_week_targets}
\alias{next_week_targets}
\title{Compute Next Week Targets for Prediction}
\usage{
next_week_targets(fit_stan_data, horizon = 1L)
}
\arguments{
\item{fit_stan_data}{Training data list used in fit}

\item{horizon}{Integer number of weeks ahead (default 1)}
}
\value{
Integer vector of global week_idx targets
}
\description{
Computes which week indices to predict based on the last fitted week.
Used to determine target weeks for out-of-sample predictions.
}
\seealso{
Other data-preparation: 
\code{\link{prepare_gq_data}()},
\code{\link{prepare_schedule_indices}()},
\code{\link{prepare_stan_data}()},
\code{\link{roll_forward_fit_stan_data}()}
}
\concept{data-preparation}
