% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nflendzoneModel-package.R
\docType{package}
\name{nflendzoneModel-package}
\alias{nflendzoneModel}
\alias{nflendzoneModel-package}
\title{nflendzoneModel: NFL Bayesian Modeling Package}
\description{
NFL Bayesian state-space models for team strength estimation
  and game prediction using pre-compiled Stan models.
}
\author{
\strong{Maintainer}: Tyler Pollard \email{tylerpollard410@gmail.com}

}
\concept{help}
