% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{print.nflendzoneFit}
\alias{print.nflendzoneFit}
\title{Print Method for nflendzoneFit Objects}
\usage{
\method{print}{nflendzoneFit}(x, ...)
}
\arguments{
\item{x}{An nflendzoneFit object}

\item{...}{Additional arguments (unused)}
}
\description{
Print Method for nflendzoneFit Objects
}
