% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_utils.R
\name{print_model_code}
\alias{print_model_code}
\title{Print Stan model code}
\usage{
print_model_code(model, package = "nflendzoneModel")
}
\arguments{
\item{model}{Name of the model (without .stan extension).}

\item{package}{Name of the package, defaults to "nflendzoneModel".}
}
\description{
Prints the Stan code for a specific model in the package.
}
\examples{
\dontrun{
  print_model_code("team_strength_fit")
}
}
\seealso{
Other package-utils: 
\code{\link{list_package_models}()}
}
\concept{package-utils}
