% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{roll_forward_fit_stan_data}
\alias{roll_forward_fit_stan_data}
\title{Roll Forward Fit Stan Data by Weeks}
\usage{
roll_forward_fit_stan_data(fit_stan_data, schedule_df, weeks_ahead = 1L)
}
\arguments{
\item{fit_stan_data}{Current Stan data list}

\item{schedule_df}{Full schedule with indices from prepare_schedule_indices()}

\item{weeks_ahead}{Number of weeks to extend}
}
\value{
New Stan data list extended by weeks_ahead
}
\description{
Extends Stan fit data by including additional weeks from the schedule.
Used for sequential updating when new game results become available.
}
\seealso{
Other data-preparation: 
\code{\link{next_week_targets}()},
\code{\link{prepare_gq_data}()},
\code{\link{prepare_schedule_indices}()},
\code{\link{prepare_stan_data}()}
}
\concept{data-preparation}
