% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{summary.nflendzoneFit}
\alias{summary.nflendzoneFit}
\title{Summary Method for nflendzoneFit Objects}
\usage{
\method{summary}{nflendzoneFit}(object, ...)
}
\arguments{
\item{object}{An nflendzoneFit object}

\item{...}{Additional arguments (unused)}
}
\description{
Summary Method for nflendzoneFit Objects
}
