% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recursive_utils.R
\name{transition_season_priors}
\alias{transition_season_priors}
\title{Transition State Across Seasons}
\usage{
transition_season_priors(priors, fit_full)
}
\arguments{
\item{priors}{List of priors (means/sds) from the end of previous season.}

\item{fit_full}{Full model fit to extract Seasonal AR parameters.}
}
\value{
Updated list of priors.
}
\description{
Applies the Season-to-Season AR(1) transition to the means and SDs.
Used when moving from Week 22 (Super Bowl) to Week 1 of next season.
}
