% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_strength.R
\name{generate_team_predictions}
\alias{generate_team_predictions}
\title{Generate predictions from fitted model}
\usage{
generate_team_predictions(fit, gq_data, ...)
}
\arguments{
\item{fit}{A CmdStanMCMC fit object from fit_team_strength_model().}

\item{gq_data}{List of data for GQ model (includes OOS games and future metadata).}

\item{...}{Named arguments to the `generate_quantities()` method.}
}
\value{
A CmdStanGQ object with predictions.
}
\description{
Generate predictions using the team strength GQ model.
}
\examples{
\dontrun{
  if (instantiate::stan_cmdstan_exists()) {
    gq <- generate_team_predictions(
      fit = my_fit,
      gq_data = my_gq_data,
      parallel_chains = 4
    )
  }
}
}
\seealso{
Other models: 
\code{\link{fit_team_strength_model}()}
}
\concept{models}
